WIP: moved checking memory exec rights to MCI
